AI-project/backend/requirements.txt
# Backend Dependencies - Flask Application pour Analyse Pétrolière

# Framework Flask
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Base de données
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Migrate==4.0.5

# Validation et sérialisation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Analyse de données (versions avec wheels pre-compilés)
pandas>=2.2.0
numpy>=2.0.0

# Utilitaires
python-dotenv==1.0.0
Werkzeug==3.0.1

